Fixed some warnings, prepared glad to be added as submodule once moved back into main branch.

+++ OpenGL-Rendering-Engine/Exalted/src/Platform/OpenGL/OpenGLRenderer.cpp
@@ -41,7 +41,7 @@ namespace Exalted
 
 	void OpenGLRendererAPI::SetClearStencil(const float stencilValue)
 	{
-		glClearStencil(stencilValue);
+		glClearStencil(static_cast<int>(stencilValue));
 	}
 
 	void OpenGLRendererAPI::Clear()
